Jam version

+++ Assets/Scripts/Enemies/Projectile.cs
@@ -24,13 +24,13 @@ public class Projectile : MonoBehaviour
 
     private void OnTriggerEnter2D(Collider2D collision)
     {
-        if (collision.attachedRigidbody == null)
-            return;
-        if (collision.attachedRigidbody.transform == PlayerState.Instance.transform)
+        if (collision.attachedRigidbody != null && collision.attachedRigidbody.transform == PlayerState.Instance.transform)
         {
             PlayerDeath.Instance.Kill();
         }
 
+        if (collision.gameObject.layer == 8)
+            return;
         Explode();
     }
